feat: 36397 - refactor structure
