+++ src/modules/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
-import { PrismaModule } from '../prisma/prisma.module';
+import { PrismaModule } from '../../shared/prisma/prisma.module';
 
 @Module({
   controllers: [UsersController],
